Fixed a small bug introduced by dateRange().
The addition of dateRange() caused the removal of a line in
getDateFromRequest() (the increment in endDate to make it inclusive in
the period), so the reports for flush and crop were sort of breaking by
not including the end date in the report.
The addition of methods to the Flush model is independent of this
change.