This is turning into a nightmare I am really strugling for a nice way to sort a table by a column.
Have implemented a pretty hacked up method but this is not work-
ing. The dictionary keys are sorted correctly as they go in but
there is no reflection of that change in the output.

Fixed a bug in the bundySignOnOff process where it would always
try sign off picker number 1.

Added sort to the GET for the all picker report.