Finished writing a prototype scale simulation app.
scale.py now represents a real scale that writes to a serial port and
makes another port available for reading. It provides a simple GTK window
with a slider to represent the weight on the scale.
This app will be useful in unit testing Gleba without having a real hardware
scale.