gui.py: Removed unneeded method main().
This method was unnecessary as its only function was to call gtk.main(),
which really should be left to the caller of MainWindow() anyway. When
we run gui.py as main, we have control over this, so we can really
replace MainWindow.main() by gtk.main().
Other stylistic issues were also resolved.