gui.py: Changed if statement to a safer comparison.
When using get_active() in a combo box, a -1 can be returned if
there is no active item. However, it's a safer practice to use
'less than' as opposed to 'equality'.
Some parameters were named in a lousy way (e.g. 'whatever', 'x'),
so now I've renamed them as 'button' or 'widget' so that they are
more informative.