gui.py: Big refactoring of add_initial_data().
When adding buttons to the GUI, this functions was wasting CPU and
declaring variables that were not really used for anything. I've
optimized lots of these things.
I've also added a layer of abstraction which allows the end-user
to configure the GUI via config.py. I've added two settings:
PICKER_COLS and VARIETY_COLS that can be used to adjust how many
columns each grid will have, leading to a much improved interaction.