gui.py: Migrate from PyGTK to PyGObject.
This is a big commit that changes all the PyGTK calls to the new
PyGObject. I've done some small tests, but there might still have
a lot of broken things. The code is still a bit smelly and I did
not change any behavior, except when things were breaking (e.g.
when trying to get an integer from a TreePath selection).

In the future a lot of the direct calls to change the GUI state
must be changed to a more elegant evented style.
I've fixed a small bug that prevented the batch to be changed.