gui.py: Removed three 'globals'.
The self.edit_*_combo in open_edit_window() were removed for safer
local variables and passed to the callback via parameters. Not
only this is cleaner (more readable), but it's also safer (e.g. no
other function will mess with this data) and it does not pollute
the 'global' (the instance variable) namespace.
The thread target count_up() was renamed to consumer_thread(). I
never understood the reason it was called 'count_up()' and
'count_in()' (recently there was a function with such name, but
now it's been removed).