templates: Removed dependency on base.html.
base.html was copied from the Django admin and customized. This
gaves us a very poor way to upgrade as the Django admin interface
modified its base.html and we were running on an older interface.
It is also very bad for bandwidth as we kinda have to copy all the
admin templates to our template directory as opposed to simply
symlinking to the Django distribution. I'll eventually remove all
the admin files from our template directory and leave it as clean
as possible.