templates/js: cleanup and symlinking.
I've cleaned up a bit the templates. For example, in HTML5 we do not need to
specify the 'type' attribute in <script> and <link rel="stylesheet"> anymore
because the specification now assumes javascript and CSS respectively.
We'll maintain the static javascript and css files with symlinks as opposed
to hardcoding their path. This makes the maintenance a lot easier than
changing paths in lots of templates. Now if we need to update jQuery or
jQuery UI, we simply replace the old scripts by the new ones and symlink
them, while maintaining the same symlink name (which is the one that templates
will use). The same will happen to CSS files.