Replaced GnuPlot by jqPlot.
Instead of relying on the server for generating graphs with GnuPlot, we now
have a much more flexible, scalable and fun way of generating graphs for the
end-user. We now use the client's processing power to generate on-demand plots
that can ultimately be altered via AJAX calls, allowing a much more dynamic
website.
The previous GnuPlot view is still in views.py. We can use it if we want as a
fallback in case the client does not have javascript enabled.