Fixed getDateFromRequest().
This utility function was breaking in some rare cases. I've refactored it and
also removed the try/except clause that not only is expensive but it also
hides some errors when it should not.
I've created a custom javascript file to contain some utility functions for
reporting. This saves a bit of bandwidth and also makes report.html a lot
cleaner.
Flush reports now also use jqPlot instead of GnuPlot.