report_all_pickers now is rendered with jqGrid.
The requests from the report page for all pickers was hitting the server
every time the user wanted to reorder the table. Now this is done on the
client with javascript (i.e. with jQuery + jqGrid). I've changed the
format of the data to JSON.
Probably the other tables will follow the same thing, even if it's just
for rendering something that looks good out of the box.