Picker report now uses jqGrid.
Instead of using a table, we now use jqGrid to display the set of values to
be plotted.
There are still some CSS quirks I don't know yet how to solve. Every time I
turn on 'display: inline;', both jqGrid and jqPlot break. I'll investigate
this later, at the moment they are both displayed as block elements.
Another quirk is that if I put jqPlot before jqGrid, the grid is displayed
as a horrendous 'tall' element, covering the graph.