glebaAdmin: Moved old Gnuplot views to another file.
Instead of explicitly import Gnuplot in views.py, it's better to put all
those old views in a separate file, and if necessary change urls.py to
point to the alternate views if the user wants it that way. In this way
we can isolate views.py from changes in the deployment (that really should
be up to the end-user if the processing should be done on the server or on
the client).