frontend: fixed missing dataset in Firefox and IE and added audio.
Firefox and IE do not support the dataset attribute which greatly simplifies
the storage of some "member attributes" of some objects. I have provided a
fallback using the more verbose getAttribute()/setAttribute().
I have also added audio to the buttons and the add_box() method. It is working
beautifully in Opera 11.10, not so great in Firefox 4.0.1 and it's sort of
breaking in Chrome 11.0.696.71 (86024), all running on Linux. Chrome refuses
to play the shorter sample 'button.ogg', even when audio controls are enabled.
Firefox has some weird delays and seems to be 'losing' some play() calls. I
hope these things are fixed in future versions.