HTML GUI: Added edit_callback() and cancel_edit().
The 'Edit' button now fires edit_callback(), which prepares an entry in the
history_table to be edited by the buttons on the interface. This avoids another
screen on the app, imho greatly increasing the speed and usability of the GUI.
The cancel_edit() callback cancels out the preparation done by edit_callback(),
restoring the normal state of operation. The only difference is that the
previously selected row is not selected after cancelling (Note: that is the
behaviour I want. If the opposite effect is desired, then we can simply 'save'
the current 'editable' row to be restored at the end).
I've also changed data-selected attributes to an additional class in the table
rows of history_table. This makes it easier for styling (although the styles
were properly applied before), but it's mainly more semantic.