HTML GUI: Added server for frontend.
The architecture for the HTML frontend is now pretty much done. An HTTP server
will run as a 'daemon' serving content for HTML clients (i.e. browsers).
I've tried serving a dummy weight from this server (gui.py) and the
client can easily issue an AJAX call requesting it. The opposite is also
easily doable by posting a request (i.e. an HTTP POST), for example for
submitting the collected weights to the server. We have an option here
to post directly to the main Django server or to gui.py, which in turn
can post to the server. I haven't decided yet which approach will be the
default. Notice that there is a problem with cross domain ajax call in
slightly older browsers, so maybe consolidating in gui.py might be more
sensible.