HTML GUI: Integrated utils.py into gui.py.
Now we have a "complete" web app that GETs the value from an http server which
in turn gets its value from utils.py which gets its value from a
scale (or a simulation app). There are some weird things such as the
value not updating, but that could be an issue with the scale simulation
app instead of utils.py. The HTTP server (gui.py) and HTML/js are
working beautifully though. =)