HTML GUI: Added a simple window.onload event.
Even though this event is not considered to be a good practice in
normal js development, that's fine for this purpose as the page is
very small, and there are no big images to be loaded. Plus, if hosted on
a LAN deployment, the page will be loaded super quickly. And to be
honest I don't want this HTML GUI to depend on other external libraries.