Web GUI: Added timestamp to boxes.
Before the timestamp was a placeholder in the code, but now it's the
time that a box is added. The function used is pretty much taken from
the Mozilla docs.
I've also removed the change in the timestamp when editing an entry. I
think it makes more sense to keep the original weighing time as opposed
to the edit time.