Added template root dir to TEMPLATE_DIRS.
The root of the template directory was not included in the list
of directories where Django would search templates. Because all the
templates were currently in that folder, the application was breaking.
Eventually we'll move each template to a specific folder, but at the
moment we better just include the root. Plus there are templates that
are applicable to the whole project.