Fixed broken {%url%} in bundy.html.
Two url tags were not being properly resolved because the old
glebaAdmin was used instead of the new gleba.apps.
The inline script in bundy.html is now a separate file so that the
client doesn't need to redownload it all the time.
Another change is that the functions in this separate javascript
file are namespaced so that they won't conflict with anything else
on the system.