Major refactor of CSV views.
I've moved out the csv related views to the report module, cleaning
even more the deprecated glebaAdmin.
The view now is totally independent of the underlying filesystem as
it doesn't write to a physical file anymore, making the process a
whole lot faster, even for the user who now automatically receives
an attachment without further clicks.