csv: error message when date in incorrect format.
Before the system was silently returning a CSV when the dates were
wrong. Now, it will redirect the user back to the page with a small
error message indicating which date had the wrong format.
When redirecting back to the page, the user is not forced to type
again the dates because they are placed back into the input field,
possibly minimizing typing.