frontend: Added JSON format to utils picker list.
Now utils can properly fetch and forward a JSON object to a
frontend. The WebGUI now defalts to the json format when fetching
the picker list, which not only is more JS like, but it's also
faster than XML and more open than our previous text format.
I have also removed a potentially dangerous eval() in gui.js. We
now use the more restrictive JSON.parse() method.
The batch list and variety list will also be converted to this form.