gleba.apps.weigh views refactored
The non-standard format that was used originally to transport data
between Django and the client has been removed.

There are two options for data markup 'json' | 'xml'. Both are
accessed from the same view.

Camel case is going to replaced with a new convention of all names
in lower case with words seperated by _ (underscore).