scale.py: Improved performance and removed dependency.
The performance is improved by removing unnecessary sleep() calls
and using the natural block and timeouts provided by Queue in
multiprocessing. Not only this makes the code cleaner, but it's
also more efficient and predictable.
I've reduced the Queue depth to a maximum of 1, so that only the
last from the slider is considered, as opposed to the previous
unbounded Queue. The previous behaviour registered all the values,
giving a "lag", which was the time necessary to consume all those
things. The behavior now is more "realtime", in which some values
"in between" are discarded. This also improves the performance and
memory usage.
There's a very strange thing in the Queue class, or maybe I do not
understand it very well. Look at the line "print ''" added in this
commit. Without this strange line (or any print), Queue was putting()
and getting() the "previous" value, like an stale value or something.
With this line there, it works all fine. Something weird that we
might need to investigate further in case this problem occurs again.