utils.py: Refactored main threading loop.
The previous loop was essentially a 'while True:' loop, and now it's
based on an independent condition variable that monitors the state
of the serial port (open/closed). This avoid a race condition when
terminating the frontend in which the client (e.g. gui-gtk3.py) asks
for the termination of the thread that consumes the serial port, but
the thread still executes a few iterations of the loop when the serial
port has already been closed. Now this bug won't occur anymore as the
condition is closely related to the reading operation and PySerial
should maintain its integrity.