GUI GTK3: Improved consistency.
These new variable names will better reflect the structure of the GUI.
Also a VBox was added to hold the picker frame to be consistent with
the left hand side (the part containing the batches, messages, weights
and offset). It also fixes an ugly visual mismatch between Frames and
VBoxes.