GUI GTK3: Added skeleton for XML interface.
The current code for the GTK GUI is very messy because of the inline
instantiation of GTK widgets. This can be improved by separating this
task to a separate XML file that we can load at runtime. We'll be
using the new GTK Builder object to create the interface instead of
the now deprecated glade library. I have created the XML by hand but
I've read somewhere that the Glade program can save in the new Builder
format. I've done it in this way not only to understand this system but
also to maintain a tighter control of what goes in the XML file.
Together with this commit there's an auxiliar program to test the
interface. Eventually it should replace gui-gtk3.py (or renamed).