utils and GUI: Added method to access DB.
The method get_active_*_xml() is not really returning an XML file
but a Python list. This is confusing and I don't like the hardcoded
indices to indicate the data that we want to extract (e.g. picker[0]
for the picker ID). It's much more readable when we use names for
the properties (e.g. picker['id']). To do this I've added very small
methods in utils.DBAPI to basically convert JSON objects to Python
lists of dictionaries (easily done with json.loads).
These changes are reflected in the new Builder interface based GUI,
but they can be incorporated in the old gui-gtk3.py file as well
even though we are planning to phase it out.