GUI GTK3: Refactored the changes of state.
Now changing states is a bit cleaner. All the assignments to state
are done in a single function instead of all over the place. The
main threading loop also does not assign to the current state and
also doesn't call set_status_feedback() anymore, reducing the size
and complexity of that messy loop.