GUI GTK3: Small optimizations.
There was something funny going on with the current_batch being
updated at every iteration of the threading loop. Now it's updated
only when we're interested in it (i.e. when state is AWAITING_BATCH),
saving some cycles in this tight loop.
I've also compressed nested two if statements into a single if for
clarity.