Added POST batch processing for list of boxes.
We were accumulating boxes in the GUI but when we committed the
operation we were still sending multiple GET requests. Not only this
is less efficient, but it's also bad in terms of semantics as we
were using a GET request to POST data to the backend (remember that
GET should be idempotent). The new method add_boxes() prepares the
data (a list of dictionaries) as a JSON object and the backend's
new function with the same name simply builds a Python list again
from the JSON object, making the processing quite similar to what
we already had.