GUI GTK3: Fixed small thing in two callbacks.
Everytime a picker_callback() or variety_callback() occurred we
were assigning to the respective self.current_* variable. Now we
only do this when we are expecting these values to change (i.e.
when the current_state is either AWAITING_PICKER or AWAITING_VARIETY).
I've also added a 'reset' to these values when the box is removed.