GUI GTK3: Renamed confusing method.
The method "history_callback(self, button)" was not really a callback
but a direct function call, and the button argument was absolutely
useless (it's not a callback triggered by any widget). To make things
worse the other function "modify_history_callback()" is named in a
similar way, adding much confusion to the structure of this GUI.
The method now is clearly named "save_box()" and it accepts no
arguments. It merely saves the current data (picker, batch etc) to
a new entry in history_entries (the source of our commit) and
history_store (part of the widget used to display these things).