GUI GTK3: History now contains multiple columns.
The history list was a simple list of strings, that was a bit messy.
Now, it's a pretty neat "matrix" that can be customized in many
different ways. It added a bit of code to add the widgets, but when
we migrate to GTK Builder, it'll be a lot nicer.