GUI GTK3: Removed all set_size_request().
These calls were mainly introduced to patch up the appearance of
the GUI in a crappy 800x600 CRT monitor when we had to launch Gleba.
We will find a better way to represent this and make the application
look better, but at the moment I'm more interested in cleaning up
the code and making the interface more maleable.