frontend: Initial migration to GTK Builder.
The new GTK Builder based frontend seems to be working fine albeit
some missing features (the edit window has not been migrated yet).
It looks cleaner and easier to read than the hardcoded UI in the
alternative gui-gtk3.py.
We will probably continue development from here and abandon both
frontend/gui.py and frontend/gui-gtk3.py. This new file is cleaner
from the start, and will allow this frontend to be better developed
in the future.