frontend: Added editing capabilities.
The editing was not migrated before, and it's a necessary part of
the Gleba. The edit window now is slightly bit more efficient, only
adding the combo boxes data once instead of every time as it was
before. It's also a bit more evented instead of directly passing
parameters to the callbacks (e.g. the callbacks will query the
GUI for the current selected row in the history list).