models: Compressed get_time_worked() method.
The get_time_worked() method was split in two methods, but now we
have a unified method that takes care of both situations.
I've also added a unit test to check this behavior. However, this
test is time sensitive in relation to the time of the day that it
runs. It needs to be run before 22:00 otherwise a Bundy entry
might be considered to be counted as the next day's bundy, giving
wrong results. The 22:00 is really 24:00 - 02:00 from the unit
test I've written that needs to write random intervals that fall
between 1 and 120 minutes.