config/utils: Removed unnecessary stuff.
The DBAPI class had a "constant" http_address which just pointed to
config.py's django_http_address. Now, this constant is called
DJANGO_HTTP_URL which gives a more concrete idea that we are dealing
with a URL, and it also uses the convention of using all capitals for
constants.
The #! line was removed from both files because they are not intended
to be run as a separate script, but to be part of a main program.
Some comments were also added.