licenses: Added more precise copyright notice.
We had first line with the copyright, year and authors, but also a
"duplicate" a little bit down below. We now have two copyright
notices, one for the period that all four of us were developing,
and a second one that started when Gleba was open sourced.
Some files had their copyright notice missing. It's fixed now.