add_boxes() now requires login.
Without a login, anyone can add boxes to Gleba by simply making a
POST request to the correct address. This makes it impossible for
deploying Gleba on a public server.
The other requests (e.g. picker_list.json, batch_list.xml etc) are
a bit less important because they do not alter the state of the
backend (they only retrieve data), but we can certainly restrict
them to users that are logged in.