Web GUI: Small refactoring.
I don't like the sudden returns in the if/else statements. Everything
now is inside this big if/elif/else block, so there's no need to
return anything. Plus, there was a small "bug" because the Python
interpreter thought req_file was undeclared in its use after the
if/elif/else block. Even though these tail after the if/elif/else
statement would only be executed when req_file was defined, it's
definitely a bad practice, and that's one of the reasons I'm removing
it.