scale.py: Added constant CYCLE_TIME.
Instead of hardcoding the period for reading/writing from/to
serial ports, we now have a better "constant" at the top-level. I put
the word constant in double quotes because Python doesn't have a real
constant as in C/C++ "const", but still we get the same effect if nobody
changes its value.