gui.py: Changed misleading comparison.
The comparison in the consumer thread was comparing the current color
of the weight label to green to check if it was in the "REMOVE_BOX" state.
Even though this works (i.e. at the moment the color of the weight label
is green only when it's in the "REMOVE_BOX" state), this is less than
ideal as that main consumer thread should not be concerned about colors
or any visual aspects.